ISDB-266: Fixed test cases
