+++ src/functionalTest/resources/responses/bulk-scan-a59-validation-output.json
@@ -1,5 +1,5 @@
 {
-  "status":"ERRORS",
+  "status":"SUCCESS",
   "warnings":{
     "items":[
