+++ src/main/java/uk/gov/hmcts/reform/bulkscan/services/BulkScanA59Service.java
@@ -74,8 +74,8 @@ public class BulkScanA59Service implements BulkScanService {
                                validationConfig.getOptionalFields());
 
         Map<String, Object> populatedMap = (Map<String, Object>) BulkScanTransformHelper
-            .transformToCaseData(transformConfigManager.getTransformationConfig(FormType.A59)
-                                     .getCaseDataFields(), inputFieldsMap);
+            .transformToCaseData(new HashMap<>(transformConfigManager.getTransformationConfig(FormType.A59)
+                                     .getCaseDataFields()), inputFieldsMap);
 
         populatedMap.put(SCAN_DOCUMENTS, transformScanDocuments(bulkScanTransformationRequest));
